# contracts/openapi.yaml
...
openapi: "3.1.0"
info:
  title: "材料实验助手 API"
  version: "1.0.0"
  description: |-
    当前后端实现的接口清单；维护依据为 backend 中的 Controller、Service、响应模型和安全配置。
    普通成功响应使用 data 和 request_id；分页另有 meta。空 data 按当前 NON_NULL 配置省略。
    204 和文件正文不使用成功信封。错误返回 application/json，包含业务 Problem Details 兼容对象、精简安全错误或内部错误；不能假定所有错误都有 detail、instance 或 request_id。

    前端使用同源 Session（默认 Cookie 为 JSESSIONID）。先获取 /auth/csrf 并保留 csrftoken，写请求携带匹配的 X-CSRFToken；登录也受 CSRF 校验。当前 SecurityConfig 另启用了 HTTP Basic。
    项目和实验读取按当前组织限定数据范围，普通用户另受项目、负责人或实验参与关系约束。创建实验或变更实验关联项目时先校验项目属于当前组织；超级管理员仅跳过项目管理成员检查。平台管理员仅能开通与查看组织目录，不获得其他组织的业务数据读取权。项目和实验受版本保护的操作使用 ETag / If-Match；当前没有 Idempotency-Key 去重实现。

    时间字段沿用现有 API 的 ISO 8601 表示，带时间的响应为含偏移量的日期时间。规范要求客户端展示及人工记录统一到 Asia/Shanghai，但当前页面部分格式化使用浏览器默认时区，尚未统一。项目和实验计划时间省略偏移量时服务端按 Asia/Shanghai 解释；调用方仍应优先提交含偏移量的日期时间。

    能力边界：ready 检查应用启动就绪及数据库连接；项目文档和实验附件正文保存在 PostgreSQL，文档 preview 返回原件。RustFS 存储和 LibreOffice 服务端转换尚未接入。
    实验参与人可随创建或 PATCH 写入；项目成员仅可读取，创建时加入创建者。开发初始化不能替代生产组织、角色和管理员引导；本契约不表示已完成生产部署验收。
servers:
  -
    url: "/api/v1"
    description: "同源 API"
...
security:
  -
    cookieAuth: []
  -
    basicAuth: []
paths:
...
  "/auth/invitations":
    get:
      operationId: "listRegistrationInvitations"
      tags: ["Auth"]
      summary: "查询当前组织的邀请码元数据"
      description: "仅超级管理员。邀请码明文仅在创建响应中返回一次，列表不包含明文或哈希。"
      responses:
        "200":
          description: "最近 100 条邀请码元数据"
          content:
            "application/json":
              schema:
                allOf:
                  - "$ref": "#/components/schemas/Envelope"
                  - type: "object"
                    properties:
                      data:
                        type: "array"
                        items:
                          "$ref": "#/components/schemas/RegistrationInvitation"
        default:
          "$ref": "#/components/responses/ApiError"
    post:
      operationId: "createRegistrationInvitation"
      tags: ["Auth"]
      summary: "签发绑定角色的一次性邀请码"
      description: "仅超级管理员。角色不能是 super_admin；邀请码有效期为 1 至 720 小时，明文只随本次响应返回。"
      requestBody:
        required: true
        content:
          "application/json":
            schema:
              "$ref": "#/components/schemas/RegistrationInvitationCreate"
      responses:
        "201":
          description: "邀请码签发成功"
          content:
            "application/json":
              schema:
                allOf:
                  - "$ref": "#/components/schemas/Envelope"
                  - type: "object"
                    properties:
                      data:
                        "$ref": "#/components/schemas/RegistrationInvitationIssue"
        default:
          "$ref": "#/components/responses/ApiError"
  "/auth/invitations/{invitationId}/revoke":
    post:
      operationId: "revokeRegistrationInvitation"
      tags: ["Auth"]
      summary: "撤销尚未使用的邀请码"
      description: "仅超级管理员；邀请码已使用、已撤销或不存在时返回 404。"
      parameters:
        -
          name: "invitationId"
          in: "path"
          required: true
          schema:
            type: "string"
            format: "uuid"
      responses:
        "204":
          description: "邀请码已撤销"
        default:
          "$ref": "#/components/responses/ApiError"
...
components:
  securitySchemes:
    cookieAuth:
      type: "apiKey"
      in: "cookie"
      name: "JSESSIONID"
      description: "Servlet Session 默认 Cookie 名；服务器配置可覆盖"
    basicAuth:
      type: "http"
      scheme: "basic"
      description: "当前 SecurityConfig 启用的 HTTP Basic；写请求仍受 CSRF 约束"
    csrfCookie:
      type: "apiKey"
      in: "cookie"
      name: "csrftoken"
      description: "先调用 GET /auth/csrf 获取并保留"
    csrfHeader:
      type: "apiKey"
      in: "header"
      name: "X-CSRFToken"
      description: "与当前 csrftoken Cookie 匹配的令牌"
...
  responses:
    ApiError:
      description: "业务或参数错误为 Problem Details 兼容对象；安全过滤器返回 status/code/detail；未分类异常返回内部错误。通常 401 为未认证、403 为权限或 CSRF 失败、404 为资源不可见或不存在。当前媒体类型为 application/json，不能假定所有非法输入均为 400。"
      content:
        "application/json":
          schema:
            anyOf:
              -
                "$ref": "#/components/schemas/ProblemDetail"
              -
                "$ref": "#/components/schemas/SecurityError"
              -
                "$ref": "#/components/schemas/InternalError"
...
  schemas:
    Envelope:
      type: "object"
      required: ["request_id"]
      properties:
        request_id:
          type: "string"
          description: "当前响应生成的标识"
          format: "uuid"
      description: "非空业务数据通过 data 返回；空 data 按当前 NON_NULL 配置省略。"
...
    ProblemDetail:
      type: "object"
      required: ["type","title","status","code","detail","instance","request_id"]
      properties:
        type:
          type: "string"
          const: "about:blank"
        title:
          type: "string"
          description: "HTTP 状态英文名称"
        status:
          type: "integer"
        code:
          type: "string"
          description: "业务错误代码，例如 version_conflict"
        detail:
          type: "string"
          description: "业务错误或校验说明"
        instance:
          type: "string"
          description: "请求 URI"
        request_id:
          type: "string"
          format: "uuid"
    SecurityError:
      type: "object"
      required: ["status","code","detail"]
      properties:
        status:
          type: "integer"
          enum: [401,403]
        code:
          type: "string"
          enum: ["authentication_required","permission_denied"]
        detail:
          type: "string"
          const: "认证或权限校验失败"
      description: "安全过滤器直接返回，不包含 request_id、type 或 instance。"
    InternalError:
      type: "object"
      required: ["type","title","status","code","request_id"]
      properties:
        type:
          type: "string"
          const: "about:blank"
        title:
          type: "string"
          const: "服务器内部错误"
        status:
          type: "integer"
          const: 500
        code:
          type: "string"
          const: "internal_error"
        request_id:
          type: "string"
          format: "uuid"
      description: "未分类异常的兜底响应，不包含 detail 或 instance。"
...
    RegistrationInvitationCreate:
      type: "object"
      required: ["role_code"]
      properties:
        role_code:
          type: "string"
          description: "当前组织有效角色代码，不能为 super_admin"
        valid_for_hours:
          type: "integer"
          minimum: 1
          maximum: 720
          default: 168
    RegistrationInvitationIssue:
      type: "object"
      required: ["invitation_code", "role_code", "expires_at"]
      properties:
        invitation_code:
          type: "string"
          description: "仅本次签发响应包含；不得记录或再次查询"
        role_code:
          type: "string"
        expires_at:
          type: "string"
          format: "date-time"
    RegistrationInvitation:
      type: "object"
      required: ["id", "role_code", "role_name", "status", "expires_at", "created_at"]
      properties:
        id:
          type: "string"
          format: "uuid"
        role_code:
          type: "string"
        role_name:
          type: "string"
        status:
          type: "string"
          enum: ["active", "used", "revoked", "expired"]
          description: "expired 为列表读取时依据过期时刻计算的展示状态，不写入数据库"
        expires_at:
          type: "string"
          format: "date-time"
        used_at:
          type: "string"
          format: "date-time"
        created_at:
          type: "string"
          format: "date-time"
...